requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.0
faiss-cpu>=1.7.0
pandas>=1.3.0
python-terrier>=0.9.1
tqdm>=4.62.0
